remove -Wint-to-pointer-cast warning
Remove the following warning:
cast to pointer from integer of different size [-Wint-to-pointer-cast]
that is caused by casting a uint64_t to a pointer.

Reviewed-on: http://git.am.freescale.net:8181/1480
Tested-by: Stoica Cristian-B18196 <cristian.stoica@freescale.com>

+++ sec/rta/include/flib/rta/fifo_load_store_cmd.h
@@ -147,7 +147,7 @@ static inline unsigned fifo_load(struct program *program, uint32_t src,
 			uint8_t *tmp = (uint8_t *) &program->buffer[program->current_pc];
 
 			for (i = 0; i < length; i++)
-				*tmp++ = ((uint8_t *) loc)[i];
+				*tmp++ = ((uint8_t *)(uintptr_t)loc)[i];
 			program->current_pc += ((length + 3) / 4);
 		}
 	} else if (!is_seq_cmd) {
