+++ sec/rta/include/flib/rta/store_cmd.h
@@ -151,7 +151,7 @@ static inline unsigned store(struct program *program, uint64_t src,
 			uint8_t *tmp = (uint8_t *) &program->buffer[program->current_pc];
 
 			for (i = 0; i < length; i++)
-				*tmp++ = ((uint8_t *) src)[i];
+				*tmp++ = ((uint8_t *)(uintptr_t)src)[i];
 			program->current_pc += ((length + 3) / 4);
 		}
 	}
